backend/requirements.txt
Django
django-cors-headers
django-rest-knox
djangorestframework
psycopg2-binary
Pillow
pytz

